EE-6630 Split HmrcClient into smaller pieces
