+++ src/test/groovy/uk/gov/digital/ho/pttg/HmrcHateoasClientSpec.groovy
@@ -4,13 +4,14 @@ package uk.gov.digital.ho.pttg
 import org.springframework.web.client.RestTemplate
 import spock.lang.Specification
 import uk.gov.digital.ho.pttg.application.HmrcClient
+import uk.gov.digital.ho.pttg.application.HmrcHateoasClient
 import uk.gov.digital.ho.pttg.application.NinoUtils
 import uk.gov.digital.ho.pttg.application.util.NameNormalizer
 import uk.gov.digital.ho.pttg.application.util.TraversonFollower
 
 import java.time.LocalDate
 
-class HmrcClientSpec extends Specification {
+class HmrcHateoasClientSpec extends Specification {
 
     private static final LocalDate FROM_DATE = LocalDate.of(2016, 6, 21)
     private static final LocalDate TO_DATE = LocalDate.of(2016, 8, 1)
@@ -29,10 +30,10 @@ class HmrcClientSpec extends Specification {
     private NinoUtils ninoUtils = new NinoUtils()
     private TraversonFollower traversonUtils = new TraversonFollower()
 
-    public HmrcClient client
+    public HmrcHateoasClient client
 
     def setup() {
-        client = new HmrcClient(mockRestTemplate, ninoUtils, traversonUtils, mockNameNormalizer, HMRC_API_VERSION, HMRC_BASE_URL)
+        client = new HmrcHateoasClient(mockRestTemplate, ninoUtils, traversonUtils, mockNameNormalizer, HMRC_API_VERSION, HMRC_BASE_URL)
     }
 
     def 'should retain any returned query params from absolute url'() {
